webserver: separated submodule setup
