+++ services/web/server/src/simcore_service_webserver/activity/__init__.py
@@ -1,1 +1,1 @@
-from .plugin import setup_activity
+from .setup import setup_activity
