+++ services/web/server/src/simcore_service_webserver/director/__init__.py
@@ -1,1 +1,1 @@
-from .plugin import setup_director
+from .setup import setup_director
