+++ services/web/server/src/simcore_service_webserver/resource_manager/__init__.py
@@ -1,1 +1,1 @@
-from .plugin import setup_resource_manager
+from .setup import setup_resource_manager
